requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dataclasses==0.6
idna==3.6
numpy==1.26.4
pandas==2.2.1
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
